tell app.component to expect to receive custom event
tell app.component to expect to receive custom event, and instructs
foodlist
it to use the value of the parent property foods in its
input foodsList, placing the foods into the child
component's foodList property to be displayed. Will also tell
the <food-list> component to connect its output onFoodSelect, our
custom event emitter, to the parent method foodWasSelected. Because it
is a custom event, we receive the actual event itself, defined by the
argument $event. Then, inside of the method we can treat it as a
normal Food object, because this is what we passed in from the child
component when we emitted the event.